Rename user requests validator to forwarding validator
